Add tests, change to ints

+++ src/main/java/com/getkhaki/api/bff/domain/models/TimeBlockSummaryDm.java
@@ -16,7 +16,7 @@ public class TimeBlockSummaryDm {
     Integer meetingCount;
     Instant start;
     Instant end;
-    Long numEmployees;
-    Long numWorkdays;
-    Long totalMeetingAttendees;
+    Integer numEmployees;
+    Integer numWorkdays;
+    Integer totalMeetingAttendees;
 }
